+++ cmp/badge/Badge.js
@@ -20,15 +20,15 @@ export const [Badge, badge] = hoistCmp.withFactory({
 
     className: 'xh-badge',
 
-    render({className, intent, position = 'center', ...props}) {
+    render({className, intent, compact = false, ...props}) {
         const classes = [];
 
         if (intent) {
             classes.push(`xh-badge--intent-${intent}`);
         }
 
-        if (position === 'top' || position === 'bottom') {
-            classes.push(`xh-badge--position-${position}`);
+        if (compact) {
+            classes.push('xh-badge--compact');
         }
 
         return div({
@@ -38,8 +38,8 @@ export const [Badge, badge] = hoistCmp.withFactory({
     }
 });
 Badge.propTypes = {
-    /** Placement of badge (default center). */
-    position: PT.oneOf(['center', 'top', 'bottom']),
+    /** Sets fontsize to half that of parent element (default false). */
+    compact: PT.bool,
 
     intent: PT.oneOf(['primary', 'success', 'warning', 'danger'])
 };